Add ticket and status models
This is main 2 models we will need in the application. Tickets will
store all the information we need and status column will be used for the
kanban dashboard itself.
I've decided to create separate model for statuses, so it can be
extended with custom or additional statuses.
Configuration thought is left for the further improvements.